change wight of habit to double

+++ lib/data/sources/habits/habit_local_source.dart
@@ -3,7 +3,7 @@ import 'package:habits_flow/domain/entities/habit_entity.dart';
 abstract class HabitLocalSource {
   Future<HabitEntity> createHabit({
     required String title,
-    required int weight,
+    required double weight,
   });
 
   Future<void> deleteHabit({required String habitId}) ;
